refactor(p3_longest_substring): rename file to longest_substring.py, update import and modified main README.md
